requirements.txt
selenium==3.141.0
Appium-Python-Client==0.52
behave==1.2.6
PyHamcrest==2.0.2
allure-behave